requirements.txt
streamlit==1.31.0
pandas==2.1.0
matplotlib==3.8.0
plotly==5.18.0
numpy==1.24.4
openpyxl==3.1.2
python-dotenv==1.0.0
seaborn==0.13.0
scikit-learn==1.3.0 